Change Expr.type_switch to return callbacks rather than locl_ty
Summary:
Instead of `Expr.type_switch` returning `locl_ty`, this diff changes it to return `env -> env * branch_info`. This encapsulates the updates done to the environment as the result of refinement, making it easier to change what these updates are.

Needed for future work where we want `type_switch` to be able to update the environment for GADT-style refinements to the type parameter env.

Reviewed By: viratyosin

Differential Revision: D60680160

fbshipit-source-id: 67b26a2bff0457bd5bac7d46c610705eb683f5bd