Eliminate Neg_class by Introducing TypePredicate.IsClass
Summary:
# Context
Currently the `Tneg` type can either be a `Neg_class` or `Neg_predicate`. We eventually want to eliminate `Neg_class`, but were waiting on moving class refinements fully to the new refinement system, which is blocked on resolving how GADT-style refinements and introduction of qualified type variables are handled. After some thought I realized we can eliminate `Neg_class` sooner by introducing `TypePredicate.IsClass`, but not using it in `TypePredicate.of_ty`.

# Implementation
This diff was mostly straightforward. I added `IsClass` to `type_predicate` and changed `Tneg` to store a `type_predicate` directly, getting rid of `neg_type`. I then followed the error messages and resolved them. Care was necessary for pattern matching, and needed to move cases that handled `Neg_class` early in the pattern match of `Neg_predicate`. Again though the OCaml type checker guided me when this was necessary.

A change of note I included in this diff appears in `Typing_case_types`. When I implemented this change and ran it across WWW, I saw cases where we were concluding types were `nothing` when we didn't before. This is because `Typing_refinement` utilizes the runtime-tag based reasoning in `Typing_case_types` as opposed to `Typing_subtype.is_disjoint`. This generally leads to more precise results (see changed expect files). However this uncovered imprecision in this reasoning.

D48814039 and D48814072 added reasoning based on sealed whitelist and requirements that would replace a `InstanceOf { name, kind = Interface }` with either the union of its sealed whitelist or intersection of its requirements. While this is a safe over-approximation for the positive case, it means we under-approximate its complement. This meant that two interfaces `I1` and `I2` that have the same requirement on class `C` would both map to the same datatype of `InstanceOf { C, class }` which lead to concluding that `I1 & not I2` simplified to `nothing`.

This was easily fixed by precisely representing the data type. Instead of replacing `InstanceOf { name, kind = Interface }` if it had a sealed whitelist or requirements, I intersect it. This resolved the issue.

Because of the improved simplification of types involving negation there was one new error revealed in WWW. The fix is put up in D60767937

Reviewed By: viratyosin

Differential Revision: D60312800

fbshipit-source-id: 3e7f2896ac8cae978934c239f21d7ab4638556fa