Remove two branches from the hashing aligned loop
Summary:
The previous change to this file introduced a new conditional branch inside the aligned processing loop to deal with reading past page boundaries inappropriately.

Having that additional conditional branch is unnecessary if we juggle the load logic around to make each block responsible for loading it's own initial data.  This removes pressure on the branch predictor and reduces the loop by one instruction (12.5% in the case-insensitive code, 14.2% in the case-sensitive code).

Additionally, our loop exit conditions were unnecessarily verbose, checking for `==0` OR `<=8`. Since each block would naturally fall through to the right code on exist we can fold these into a simple `>=8` to restart the loop.

All of this results in the removal of two conditional branches from the loop body, reducing pressure on the branch predictor and improving code density by 28.5% in the case-sensitive case and 25% in the case-insensitive case.

Differential Revision: D60815868

fbshipit-source-id: 1654b5bfb6d8cfe0e26a4d4c86ab2eec383fce2c