Save an instruction inside CRC-based hash-computation loops
Summary:
D60815868 partially took care of my suggestions on D60760907. This diff takes care of the remaining suggestion, which further eliminates a `cmp` instruction inside the loops, bringing down the number of instructions inside the loop by 16-20% for the different variants.

The trick is to change the invariant for the loop so that `x1` holds the number of bytes left to be processed //after// the next loop iteration, instead of the value //prior// to that. This allows us to count down to zero instead of 8, so that we can use a `sub` instruction to also set the flags (turned into a `subs`) and avoid the `cmp`. This trick saves one instruction inside the loop at the expense of one extra instruction after the loop to adjust `x1` (the adjustment prior to entering the loop replaces a previous `cmp` instruction).

Reviewed By: mdko

Differential Revision: D60841925

fbshipit-source-id: e1c735033eaaa7cb45507782ca831bd7d664edaf