Fix smashable intercept instructions
Summary:
The smashable intercept implementation on x86_64 was swapping between a `jcc`
instruction and a `nop;jmp` pair of instuctions.  This is not safe because the
read/decoding of the `nop;jmp` pair of instructions is not atomic, e.g. a thread
may be interrupted by the OS right after the `nop`, with the PC pointing to the
`jmp` and, by the time it resumes, if the code was patched to a `jcc`, the PC
will be pointing to the middle of this instruction (which will become an `add`
with a different instruction length, see P1516397503).

This diff fixes this issue by replacing the `nop;jmp` pair with a single `jmp`
instruction with a `rex` prefix so that it has the same length as the `jcc`.

Reviewed By: ricklavoie

Differential Revision: D60872700

fbshipit-source-id: 1af82055b3145c0aea56f8960a395740538f482b