Cleanup usage of getBlock()
Summary:
When getBlock() for a SrcKey that is not part of the translation gets called,
it will produce an exit block with ReqBindJmp. While doing so, it will copy the
current frame state into the block. While ReqBindJmp is terminal and does not
really need data on anything, it marks the block as an exist block and its
postconditions are used. If getBlock() is called too early, we get an incorrect
list of postconditions, affecting region formation.

Fix this by deferring calls to getBlock() until we actually do the Jmp.

Reviewed By: mdko

Differential Revision: D60879848

fbshipit-source-id: f88f4963d5cd3990a9b159ad27c95ae9362d9d1b