Refine refine vs changed postconditions wrt loads
Summary:
Load instructions (such as LdStk and LdLoc) do not change the value. They just
refine it based on the type argument.

Do not mark the value as maybeChanged, so that if there are no other
instructions changing the value in the basic block, it will be listed as
refined rather than changed in its postconditions.

This flag is really used only in one place -- apply_transfer_function() as part
of region-prune-arcs algorithm. If we already knew a better type for the
location, it will allow this algorithm to use it and potentially prune
more unreachable arcs.

Reviewed By: ricklavoie

Differential Revision: D60801088

fbshipit-source-id: c47aa8ffac892b3ec9351c68fda367c903841803