Add GenericEnumClass to ext_enum
Summary:
The type checker bans direct reference to `HH\BuiltinEnumClass` to prevent user land classes from trying to extend this class and to hide the implementation detail of the enum class feature. Unfortunately this prevents expressing valid programs that attempt to operate generically across enum classes. An example is constraining a type constant to be some enum class.

This diff addresses this problem by adding an interface `HH\GenericEnumClass` that is implemented by `HH\BuiltinEnumClass`. To prevent users from implementing the interface, we seal it to the `HH\BuiltinEnumClass`. We only make the change to the ext_enum.php.

I don't know if there are any other changes needed besides adding the interface, so adding some folks from HHVM as reviewers just in case.

Reviewed By: viratyosin

Differential Revision: D60775509

fbshipit-source-id: 7770b0df9ed64ce9502de4bb61ddc57f39e54212