refactoring
Summary:
In `thrift.ml`, move the thrift path from the `container` type  to the `t` type.

In the context of xlang indexing, the mutable container field tells the indexer the thrift declaration of the container being indexed (when it's indexing a member method for instance). However, the thrift_path is available on a file basis (from the file header). We need this in the next diff for indexing thrift constant (no container info available but we need the thrift_path).

Reviewed By: josefs

Differential Revision: D60935937

fbshipit-source-id: a03be16310f4ff8599733e31b152069375e06096