report cycles outside of localize
Summary:
`localize` should not report errors. It's a transformation operation that can be run anytime, and should not be concerned with detecting errors. It might be run multiple times on the same input, reporting multiple times the same errors, or errors in different files, which then requires deduplication and filtering errors. It currently reports type alias cycles and a few others.

This diff moves cycle reporting outside of `localize`. We don't re-implement cycle detection (I've tried, but type constant logic is difficult). Instead, since `localize` has to detect cycles anyway to avoid infinite recursion, we lean on that. We make `localize` gather and return all the detected cycles, so we can report on them in the caller. (If we ever want to do type alias expansion out of localize, we could use the same mechanism of bubbling up cycles in the new module doing expansion).

Other benefit: it should make it easier and more principled to fix a few issues with cycle detection, e.g. with opaque types (including case types) and mutual recursion, including infinite recursions.

Review order:
- start by reading the mli of Type_expansions, then Type_expansions itself.
  - then typing_defs.mli and typing_defs.ml
- a lot of the other changes in the diff are due to API changes from Type_expansion

User-facing changes:
- when a cycle is detected, some subtype errors are silenced. This may change with future diffs.

Reviewed By: mheiber

Differential Revision: D60387807

fbshipit-source-id: 1a5f3d5883d4ecb3ab23bd32d56bfac20bdfc007