Avoid calling `union_list` when localizing the RHS of a case type
Summary:
The RHS of a case type is known to be a union of disjoint types. Knowing this, we can avoid calling into the costly `union_list`, which is normally an attempt to simplify the union after localization.

This should have perf benefit on case type code, but also prevents one kind of infinite recursion with mutually recursive case types, where the call stack would be something like subtype -> localize -> union -> subtype -> ... The next diff should prevent infinite recursion of subtyping with recursive case types altogether.

Reviewed By: mheiber

Differential Revision: D60832879

fbshipit-source-id: 31ca1cd03b45e7a136c178ebb40fafaa815e4759