Add support for thrift python to SvcThriftServer
Summary:
SvcThriftServer is an old but still used way to spawn thrift servers. It isn't too much work to get it to work with thrift python, as it just needs to have the correct fields passed to init.

Also refactoring the creation of PythonAsyncProcessorFactory since I realized that there was some code getting copied in multiple places that didn't have to be.

Reviewed By: fried

Differential Revision: D60548694

fbshipit-source-id: 53fe194d468cbc7357283c708a91aa413fa5ef71