promote from facebook directory: no secrets here
Summary:
## What

- Delete duplicated code
- Migrate non-duplicated code from 'facebook' directory

We had two files called test_lsp.py, one in a 'facebook' directory, with most of the code duplicated.

## Why

- No secrets contained within, so I don't think the 'facebook' directory was serving us well.
- Confusing, I kept getting the files mixed up
- So much duplication

## How we got here

My guess is that we were using the 'facebook' directory as a staging area for serverless_ide and then never migrated out of the facebook directory.
I think it's probably best not to do that:
- easy to forget to migrate
- it's OK, imo, if outside-worlders know that we are developing new features for Hack LSP

## Potential follow-on

Files and functions could be renamed to remove "_serverless_ide" from their names and comments could be updated, too, but I've hit my timebox long ago.

Differential Revision: D61012682

fbshipit-source-id: 4073929beb8ae0244189057c7305cdd09c494cb8