Typing_subtype.Type_switch.simplify - only move dynamic outside for false side
Summary:
Alternative to D60428944, changes the compromise made for doing a type switch on like types.

See inline comments both removed and deleted for context.

Like D60428944, this fixes a problem that shows up when turning on shape type predicates, but has the advantage that I think it's correct w/o having to check the predicate.

This also fixes a problem w/ refining `~nonnull` by `null` that andrewjkennedy noticed.

Reviewed By: dlreeves

Differential Revision: D60936561

fbshipit-source-id: e681633ad0498e0109afd7751aa0b30014311e27