Add hh_single_format and fast tests for Hack language server formatting logic
Summary:
## What

introduce a new binary for Hack team to iterate quickly on the formatting logic in the server.

## Why

I got really tired of rebuilding hh_client and opening a repo and typing around and saving.

## Details

> same as in a comment in the code

For lightweight testing of the formatting logic in **the language server**.

Usage: hh_single_ide_format $FILENAME

Logic in the language server is mainly for turning hackfmt results
into minimal edits to send to the language client (such as VSCode)

There are usually better places to test:
- hackfmt tests for hackfmt
- [lsp integration tests](https://www.internalfb.com/code/fbsource/[ea56561e576d62d671cf3173a2c1be5c0466e795]/fbcode/hphp/hack/test/integration/test_lsp.py)
   test end-to-end, including the logic that converts between LSP line+column and Ide_api_types.range

Differential Revision: D61022050

fbshipit-source-id: c07bf06aba44ddbe2285f3fe6797d2cfddb0770c