Fix decodeAndGetParam cursor bug
Summary:
Fixes a bug in decodeAndGetParam().
This function parses the OuterClientHello extension from the client hello.
Currently this function utilizes a cursor that is moved pass the outer client
hello buffer range when multiple ECHConfigs are loaded in the ECHDecrypter.
This results in an out_of_range exception being thrown.

This diff fixes this by calling getExtension once and saving the result.
Note that the RFC specifies that the client should select **one** suitable
config to use ECH (https://www.ietf.org/archive/id/draft-ietf-tls-esni-20.html#section-6.1). It is the __client's__ job to select which ech config to use
from the server's set of ECHConfigs. There should not be a case where
multiple configs are utilized by the client.

Reviewed By: mingtaoy

Differential Revision: D60994613

fbshipit-source-id: 08625fd221738e29230202c670d204e17bda35a2