Error filtering for the LSP client
Summary:
We implement error filtering for the LSP client, that is filtering out warnings for CLI flags like `-Wall`, `-Wnone`, etc. Of course these options cannot easily be passed to the LSP client, but we still need to implement the behaviour or config flag `warnings_default_all`.

All the filtering logic was already implemented in `ClientCheckStatusFilterWarnings`, we just need to call into the API, which is pretty much just these two functions:
```
module Filter : sig
  type t

  (** [default_all] controls wether an empty list of switches is equivalent to [WAll]
    as opposed to [WNone] *)
  val make :
    default_all:bool -> generated_files:Str.regexp list -> switch list -> t
end

val filter :
  Filter.t -> Errors.finalized_error list -> Errors.finalized_error list
```
The filter is created once and for all at the start of the client, and passed down to where we receive and process errors.

Reviewed By: mheiber

Differential Revision: D60967801

fbshipit-source-id: f20b72485433cc5a3b8537a8c9d6c196eb4378c1