Filter warnings on server side instead of client side when not streaming errors
Summary:
When not streaming errors, errors are bubbled up from the server to the client via RPC (which is not the case when streaming errors: in that case errors are serialized via the errors.bin file for any client to read).

On the server side of the RPC call, we already do `take_max_errors` to account for the `--max-errors` flag, so it makes sense to filter errors before we do that, and therefore to filter errors on the server side of the RPC code. This is what this diff does.

To do that, we pass the error filter via RPC to the server - see `ServerCommandTypes.STATUS`.

As a result I'm also moving clientFilterErrors.ml to client_and_server/filter_errors.ml

Reviewed By: mheiber

Differential Revision: D61125797

fbshipit-source-id: 18932578e86bfa0fefe2b09b4996a83852ad0e77