hack warnings: adapt error summary
Summary:
The "error summary" looks something like
```
10 errors found
```
and comes in the output after displaying errors.
We adapt it to also show the warning count.
We make sure that the string "No errors!" still appears if there are only warnings, because some scripts seem to rely on it.

Also make sure that the exit code is non-zero only if there are proper errors.

Reviewed By: mheiber

Differential Revision: D61126642

fbshipit-source-id: adac41a092a634c3821d65e513ffd0d8ad3f0529