Simplify enum type flattening
Summary:
I tried to reason about what `firstEnum` is doing since it really complicates the logic in type flattening. I think we don't actually need it, and all we care about is whether we are _inside_ an enum or not.

I still think there is room to simplify the case where we have a class symbol inside an enum, but can do that after.

It makes the error messages a little bit worse, but I think it makes this easier to understand.

Reviewed By: ricklavoie

Differential Revision: D60816349

fbshipit-source-id: f4670cd58c22f2bbd48cfbc0e82d18af7ebc52d2