Get rid of state mutex (try 2)
Summary:
The existing mutex (`run_state_mutex_`) protected the `cancel_on_run_` variable.  Folly::Synchronized does this in a much safer way as it become impossible to access the protected data without the mutex.

Note that the `state_` variable is partially protected by the original mutex as well.  It is partial because it only protects writes to the variable while reads are allowed without mutex protection.  The only time this is important is on the race between `run()` and `cancel()` as this is the only time this variable can be changed in different threads.  Thus the `state_` variable is only changed in `cancel()`, and `run()` when holding a lock on `cancel_on_run`.

Reviewed By: aditya-jalan

Differential Revision: D60970118

fbshipit-source-id: 9b5454987331060490c4350b90e18ea641506c17