Simplify handling of events and timeouts (try 2)
Summary:
The ConnectionSocketHandler class is a helper class that kind of abstracted the handling of event base events and timeouts.  However, the class needed the operation as a pointer to manage anything.  It would make more sense to derive Operation directly from the two classes (EventHandler and AsyncTimeout) directly.  This avoids extra indirection as well as a class of errors if the `op` (Operation) pointer never got set.

This also ends up simplifying a bunch of stuff.

Reviewed By: aditya-jalan

Differential Revision: D60970121

fbshipit-source-id: 4cc39cca4b2090d55261a7c9d23134f279d5504c