Deserialize value in container storage directly
Summary:
Create default value directly in container, and deserialize it into allocated
storage without `std::move`.

This is a behaviour change if we have key duplicates in deserialization stream.
Before, if `emplace` wasn't successful we just threw away value for this key.
Now, we are going to override value with a most recent one.

Based on the previous discussions on this topic, behaviour in case of keys
duplicated is not specified, so it's fine to change it if we have a good
reason.

Reviewed By: thedavekwon

Differential Revision: D60964997

fbshipit-source-id: 1bdc5698ef96c04b5560ac54fa23aa3036ca3f7b