Fix HHBBC handling of async memoization
Summary:
HHBBC was conflating the two cases of memoization information being
Bottom and a wait-handle containing Bottom. A wait-handle containing
Bottom was being treated as just Bottom, which meant HHBBC inferred
the memo-cache was always empty. To fix this, store the types for the
eager and non-eager cases separately. Thus, there's no ambiguity.

Reviewed By: mdko

Differential Revision: D61254673

fbshipit-source-id: a3cd797df81f4c0c2ead91f6aa469c2141cda99f