Do not present uninformative type under dynamic in hover
Summary:
If a function or method is checked both statically and under dynamic assumptions, we show two types in hover, a typically message showing `vec<int> (dynamic when called dynamically)`. If the types are identical, we show only one type.

Messages such as `~int (dynamic when called dynamically)` are not very helpful, as `~int` already includes dynamic. This diff tweaks things a bit so that if the dynamic-pass type is no worse (wrt subtyping) than the static-pass type, then don't show it.

Differential Revision: D61333600

fbshipit-source-id: 7aa8189dfc4999ee6564f7b5aac48dff22988292