union fields
Summary:
We miss thrift references for union fields

See for instance
https://www.internalfb.com/code/www/[cf90de00c18d]/flib/intern/glean/__tests__/GlassSmoke.php?lines=18

It's because the `get_union_field` methods isn't comment-annotated in the generated file. Only the first (protected) field is annotated, but it's usually not the one client will refer too.

We don't want to add more comments, but we can add more state in the thrift indexing context so we can identified those members which correspond to a thrift union field.

In this diff, we build a HackToThrift fact for all fields of the form `XXXthrift_field_name` appearing in a class generated from a Thrift union, where `thrift_field_name` is the  name of the annotated, protected field.

Reviewed By: josefs

Differential Revision: D61271387

fbshipit-source-id: d4ae808dc17492dbc77c3f62909d60f5b337061a