Error on internal accesses too
Summary:
Given the current implementation, if we had a cross-package-boundary module,
`internal` element accesses within this module would incur neither a module nor
a package visibility violation, when in fact we should be respecting the package boundary.

This diff lets us perform package check on any top-level symbol reference,
not only on symbols with the public modifier.

Reviewed By: francesco-zappa-nardelli

Differential Revision: D61226736

fbshipit-source-id: 65c4df0afe955a7ce4201ba10c17bd74bcc28747