Introduce CertificateSerialization abstraction for use in PSK serialization
Summary:
`fizz::client::serializePsk` and `fizz::client::deserializePsk` now takes an instance of a `CertificateSerialization`, which implements the serialization / deserialization logic for a given cert.

This allows PskSerializationUtils to not have a hard dependency on a concrete certificate implementation (previously, in tryWriteCert, it would use OpenSSL). The choice of the certificate implementation is put on the caller.

Reviewed By: zalecodez

Differential Revision: D60792486

fbshipit-source-id: e9ad97baf2206e548a43fac96a9ac3fdeabd8146