Implement ast-dump program for Whisker
Summary:
This diff adds a tool called `ast-dump`, which when pointed to a file (or `stdin`) will try to dump the Whisker AST in human readable form, and/or output diagnostic results.

This tool can be very helpful when debugging the compiler. For example, a successful parse can create an incorrect AST due to a bug. This tool can be used to diagnose such issues.

See the `usage()` function for a description of how this program is intended to be used.

Reviewed By: yoney

Differential Revision: D61301753

fbshipit-source-id: 6754403380f06a9d21ba8ac61876bb5f200ed765