Group Common Type Predicates under IsTag Variant
Summary:
# Context
Type predicates today are of three forms:
1. Structural, deep testing of a tuple type (i.e. `is (int, string, bool)`)
2. Structural, deep testing of a shape type (i.e. `is shape('a' => int, 'b' => string)`)
3. Shallow test by examining the runtime tag of a value (i.e. `is int`)

Currently the shallow tests are implemented as distinct variants like `IsInt`, `IsString`, or `IsBool` which means you cannot easily group these variants together in a logical grouping such as, do this if the predicate is a shallow tag test.

# Solution

This diff addresses this by introducing a `type_tag` type. A `type_tag` is representative of the runtime tag associated with a value. A `type_tag` can then form a `type_predicate` with the `IsTag` variant which is used to encode the shallow tag test. This will be used in D61362998 to clean up some code inside of typing_refinement.

Of note the `type_tag` type is similar in spirit to the `Typing_case_types.DataType.Tag.t` type, but that type is currently not exposed outside of that module. Ultimately it makes sense to combine/relate these types in some  way, but for now we focus on replacing the various `IsX` variants of `type_predicate`

Reviewed By: viratyosin

Differential Revision: D61356047

fbshipit-source-id: f32d1e87f82198e3733725d8c645cd6886c71cc2