Add construction logs to aid in debugging
Summary:
The shutdown log usually tells us which singleton's destruction is blocking.
However, it's hard to tell which singletons are pending destruction.  Having
this information helps to debug circular dependencies, where destruction of a
singleton scheduled for destruction later may be blocking the earlier
singleton's destruction.

The shutdown log persists through forks, it was hard to see which entry each
shutdown log entry corresponds to, so this diff also adds a pid to the log
lines.

Reviewed By: andriigrynenko

Differential Revision: D61286579

fbshipit-source-id: 744f10a97e9f1e2a89ca68750601698f10b044c6