Add runtime logging for speculation
Summary:
I'm trying to measure the impact of my sandbox speculation changes.
I've been looking at the translations before/after, but I think it would also be helpful to have a sense of how many times we succeed/fail from these speculations to sanity check we are taking the fast path in majority of cases.

Waiting for data to show up in scuba UI.

Reviewed By: ricklavoie

Differential Revision: D61494886

fbshipit-source-id: 637e13dae17c12c087bde319e48a583ddb71e429