Strip mstch::lambda
Summary:
In preparation for moving the rendering to Whisker, let's remove `mstch::lambda` which is not supported by Whisker. Indeed, this feature is entirely unused by Thrift!

The API of `mstch::lambda` is super weird in that it exposes the raw (unparsed) text to user code. We can achieve similar capability with `whisker::native_object`.

 killitwithfirejango

Differential Revision: D61509486

fbshipit-source-id: 30e82522a1f0717c0d0b1f078c7fec1e2669d99f