Fix soundness hole instantiating this in meth_caller
Summary:
The `this` type is interpreted unsoundly for `meth_caller` function pointers: it should not be interpreted as an "exact" type because the function pointer might be applied to an instance of a subclass. For example:
```
class C {
  public function get():this { return $this; }
}
class D extends C { }
```
The function pointer `meth_caller(C::class, 'get')` is currently given the type `(function(C):exact C)` which is unsound (consider calling it through an instance of `D`). Instead, let's give it the type `(function(C):C)`.

This is not the most precise type that could be given. In future, we should consider using first-class generic function types and give it the type
```
(function<Tthis as C>(Tthis):Tthis)
```

Differential Revision: D61593443

fbshipit-source-id: d001a2ff6b0571f3e9e5e7c7f23b4ad98122938d