S365169: Add a metric for pending upstream connections
Summary:
We want to explore setting concurrency limits on the pending upstream connection count. To do that, we need to have some visibility first to understand the normal vs problematic behavior.
This change adds a metric providing that visibility.
The challenges here are differences in behavior between different HTTP versions, different transport protocols and coro vs legacy.
On HTTP 1.1 path, client resetting a TCP connection does not trip any cancellation on the upstream happenings. So, after the configured number of retries, the upstream connection aborts, correctly calling connectError(). For QUIC, it does, and it didn't invoke any callbacks, and the timer is removed, so giving up on connecting was resulting in a counter leak. D61172051 addresses that.

I am also changing the protocol for shutting down the Prospect backend servers, allowing them to stop without needing to process a hacky inline stop signal. I need this to handle the case of blocked connection to backend. Both methods are still supported.

Reviewed By: kwaugh

Differential Revision: D40990110

fbshipit-source-id: 32d1bbb412ec45a1a5e1a336138fbb395f9afbdd