move client from daemon specific client to generic client
Summary:
I wanna log the client for all requests so we can get a better understanding of
Watchman clients. So I am gonna want the pid for all clients not just daemon
clients.

Moving the pid into the generic client class (from the daemon client class
where it is now) so I can use the pid in all cases.

Reviewed By: MichaelCuevas

Differential Revision: D61480179

fbshipit-source-id: b9f44209063bae373bf1e1f54b7354dd2f0c791c