add Union type enum
Summary:
Our codebase is plagued by Thrift Union visitors. This diff attempts to offer an alternative way to access unions and allows us to remove all these eyesores.

The idea is to have Thrift to generate a nested enum `Type` so we can exhaustively match on all possible union types.

This will allow us to write code like:

```
switch (anUnion.getThriftType()) {
  case anUnion.Type.BRANCH_1 -> System.out.println("Got Branch 1");
  case anUnion.Type.BRANCH_2 -> System.out.println("Got Branch 2");
  ...
}
```

In Kotlin:

```
when (anUnion.getThriftType()) {
  anUnion.Type.BRANCH_1 -> {
    println("Branch 1")
  }
  anUnion.Type.BRANCH_2 -> {
    println("Branch 2")
  }
}
```

When new union branch is added, the Java compiler should be able to notify the users to handle these cases with an error.

 ---

The downside is that, if you ever need to retrieve the value in the union, you will have to do it manually.

```
anUnion.Type.BRANCH_1 -> {
   val inner = anUnion.getBranch1()
}
```

It *may* be easy to make mistakes where you test for one branch and attempt to access the value of another branch.

Differential Revision: D61418545

fbshipit-source-id: e9609ed4a7871fea5604f395ee9859b027368a14