Prevent excessive copying of whisker::object
Summary:
`whisker::object` is passed around by value often in `render`. Consider a template like `{{a.b.c}}`... the `whisker::object` backing `a` & `b` are necessarily copied during `eval_context::lookup_property`. If both are large `whisker::map`s then this is immensely wasteful.

This diff changes the `native_object::lookup_property` API to return `const object&`. The rest of the changes are the transitive side-effects of this change.

This requires `native_object` implementations to ensure that the objects are kept alive for the duration of `this`. Small price to pay for the wins.

Reviewed By: vitaut

Differential Revision: D61680001

fbshipit-source-id: faf23b72ca4a35bc5032aa2f6fc6554a630be40f