Register useIntrinsicDefault annotation check in generator and modify Enum mustache file
Summary:
In diff1, synced to internal group post (View Butterfly Rule):
https://fb.workplace.com/groups/120844258624995/?post_id=1393921887983886, we added an annotation named UseIntrinsicDefault to offer an option to use a default value.

In diff4, we have not yet added the corresponding code to the generator.cc to check whether the thrift file contains the annotation, but we still generated MyUseIntrinsicDefaultEnum.java for comparison with diff5.

In diff5(This diff), we implemented the generator.cc and .mustache file.
You can check the change in MyUseIntrinsicDefaultEnum.java.

Reviewed By: lic9, avalonalex

Differential Revision: D61742298

fbshipit-source-id: 465d239d8b21db6076d6064f4f5042e85ce0068e