Optimize non-persistent Func* in jit 1/2
Summary:
Increase cases where we trust `Func*`. Previously was just when it was persistent. Its also known when its defined in the
 same unit as the callsite, since changing the callsite will reload the funcs in the unit.

Introduce `m_func` in the `NamedFunc` entity table, similar to `m_clsList` for classes.
Previously, we just stored an `rds::Link<Func>` in `NamedFunc`. When we'd go to `lookupImmutableFunc()` from the jit,
it would never match for non-persistent funcs since you can't access RDS normal from the jit. When I was dumping the
handles it looked like the generation number for the handles was 0 when I inspected them from the jit, despite them being
 1 earlier for normal handles. Is there something invalidating them that is doing this?.

Now, you can get a `Func*` in the jit using `m_func` without going through RDS.

NOTE: I discovered a bit of an issue with some of the assumptions about non-persistent `Class*` and `Func*` that we
"trust" in the jit (ie Class::lookupKnown() here https://fburl.com/code/v7ouibwx). We select the candidate that we are
 going to test whether we trust by grabbing the last symbol to be loaded for a given name (ie off the `m_clsList`. If we
 trust a `Class*` or `Func*`, its either persistent, or guaranteed to invalidate the callsite function if the `Class*/Func*`
change. Given this, grabbing the last symbol to be loaded works perfectly fine, assuming we only have one jit thread.
 However, this does _not_ work, if we have multiple threads jitting, like we now do for sandbox mode. Now, its possible
 for the last thing to be loaded (and placed on `m_clslist`) to be something from a different request with a different
`Class*` for `B`! Yikes. This is very unlikely as it would require a very precise interleaving of defining Class B differently
 in different requests that are made at similar times.

I think the solution here is to either (1) pessimize and guard all non-persistent `Class*` and `Func*` with `funcId/classId`
 checks in jitted code or (2) Add notion of a thread-local `m_clsList/m_func` for each jit thread so we can guarantee a
 `Class*/Func*` was the last symbol that a specific thread loaded (for cases where Func* is not persistent but can be
trusted ie defined in same unit as callsite).

Can guard these conditions on `Sandbox::Speculate` flag for now while I determine a fix. I think this failure mode is
super unlikely, so getting some performance numbers before figuring this out should be ok.
----
[RFC] Is there a good way programmatically prevent the jit from inspecting RDS::Normal? Last week I hit some assertions
 in `VMRegAnchor` about not finding expected fixups which was a symptom of calling from jit instead of a request. Is
 there something similar I can add to inspecting `getCachedFunc()`?

Reviewed By: jano

Differential Revision: D60261775

fbshipit-source-id: 12d23632e248af5a9846257455f46d6e4dba0551