Upgrade the thrift diagnostics to include column number, original text, replacement, and a box for additional metadata
Summary:
The current thrift diagnostics are fairly bare bones, they only supply a file, a message and a line. By adding a bit more data we can then use this to leverage various linter tools to supply auto-fixes to common thrift mistakes.

I also added a `std::any` just to hold additional metadata which is not specific to thrift but might be needed by whatever particular linter is leveraging thrift's diagnostics. Concretely in this case, `arc lint` has a key `bypassChangedLineFiltering` which is relevant for the linters in `thrift/facebook/linters` but doesn't make sense to live in the thrift compiler, so I put a std any so that this logic can be contained within the linter code.

Reviewed By: vitaut

Differential Revision: D61307052

fbshipit-source-id: 93e441ee88b4ac22b0ffc2e2958e63ee703f66c2