rename symbolrefs tests
Summary:
Context: We're getting ready to roll out static enforcement on packages v2. There are a number of places where we bypass the typechecker check to match symbolrefs behavior. For example, a cross-boundary generic would not incur a typechecker error because it'll be erased and its file will not get pulled by symbolrefs during the build.

To increase confidence on static enforcement, I will port typechecker tests to `hphp/test/slow/packages/symbol_refs/` to ensure no elided typecheck leads to a runtime fatal. This diff simply modifies the caller/callee names in existing tests to be consistent with those in the incoming tests.

Reviewed By: DavidSnider

Differential Revision: D61828625

fbshipit-source-id: bb38239d097cc616792e64d40492bfd9e5c79fa0