Allow symbolrefs union deployment for repo build
Summary:
Today if the active deployment is set, we filter out any files that are not in the active deployment from the repo build, even if ForceEnableSymbolRefs is set. This means symbolrefs outside the active deployment will not be pulled into the build either.

For intern-prod separation using deployments, we want to enable build with the active deployment + symbolrefs, hence this diff.

To summarize the end behavior:
- If no active deployment is set – build everything
- If an active deployment is specified without ForceEnableSymbolRefs – only build the deployment
- If an active deployment is specified with ForceEnableSymbolRefs – build the deployment + symbolrefs

Reviewed By: DavidSnider

Differential Revision: D61828624

fbshipit-source-id: 9fe4b5e27d2eccda39378530ee1bb7fc623403b2