Fix race in shallow RequestContext copy
Summary:
If a shallow `RequestContext` copy is done concurrently with mutations, we need to synchronize with the mutating thread to get a self-consistent state snapshot.

The race was root-caused by Gownta.

Shallow context copies are not in hot paths, so acquiring an uncontended read lock should have negligible cost, and we don't need more sophisticated solutions.

Reviewed By: Gownta

Differential Revision: D61984998

fbshipit-source-id: 64c875730a2fb56befba5f5cc7775a58317dfb79