no unique F14Chunk empty-tag-vector instance
Summary:
With `F14Table`-based containers like `F14FastMap`, we have observed the impossibility of guaranteeing a single, process-wide, globally-unique instance of the chunk empty-tag-vector.

Instead, we pick a unique invalid representation of the first pointer-sized word of the chunk and compare against it. This comes to a single extra indirection, plus the instructions for comparing against a constant which were already present anyway.

Reviewed By: akrieger

Differential Revision: D61808981

fbshipit-source-id: aa71b711f683f5f4e13880fed7b349d32f814fa5