Do not sort tuple elements when comparing types
Summary:
There is a bug in our implementation of `ty_compare` for tuple types, whereby we don't care about the order of elements; we use the same logic as for unions and intersections, which is ok there, but wrong for tuples.

But a test now fails (see commented out code in test `split_a_neg_tuple.php` dlreeves viratyosin), and I do not know why.

Reviewed By: enetsee

Differential Revision: D61983191

fbshipit-source-id: 3a0bccccfc2329d5e5e252efe0e3c8131376c622