Remove Generic from ApproxSet Module
Summary:
# Context
When I originally added support for attaching reasons to tags in D48653810, I made a change to the interface for ApproxSet that required its type `t` to take a type variable (generic) so it was `'a t`. This added complexity to the interface and is making it harder to change now that I plan on having it report relationships between sets other than if they are disjoint. This diff eliminates the need for this generic without sacrificing the quality of error reporting.

# The Change

The main reason I needed to introduce the generic was I stored a type (the origin) in each `DataTypeReason.t`. Since I wanted to use the logic for types across different phases I needed to make `DataTypeReason.t` also accept a phase generic which would be attached to each tag in the ApproxSet. Re-examining the code, I don't need to store the origin in each reason.  Instead I can store it along side the ApproxSet.

This required one minor change in `ApproxSet.disjoint`, that whenever I flip the sets I make sure to flip the result. This will guarantee that when we report `Unsat`, that `left` will always correspond to a tag in the first set and `right` a tag to the second set.

From there it was mainly cleaning up the relevant code in `typing_case_types.ml`

Reviewed By: CatherineGasnier

Differential Revision: D61954159

fbshipit-source-id: d26740d9503829105b45f8b8657e31014f20bfcd