Speculate ldCls to optimize consumers of Class* from stack.
Summary:
Was trying to figure out how to optimize `FCallClsMethodM` and found that since `FCallClsMethodM` gets its class from the stack, we need to make sure the class `SSATmp*` we get from the stack is a constant (even if its being speculated).

Optimize `ldCls` to speculate, so consumers of these values on the stack can be optimized.

Followups:
- Allow jit to load classes: I saw an example in call-soft-1.php where if we remove the call to `new A()`, the speculation does not happen since the jit doesn't have a candidate `Class*`. Need to do this

Questions:

It seems like if the `StLoc` and `LdLoc` for `$a = "A";` get split up between translations, in the second translation all we guard on is that `$a` is a Str (and we emit a `LdCls` rather than `LdClsCached` and we can't speculate at all. Can we not guard on exact values? I thought we sometimes did this for exact classes? ie `DataTypeSpecialized`

Reviewed By: ricklavoie

Differential Revision: D60557145

fbshipit-source-id: 84a837ad860900f361020e65eae9dc8e146c0a40