Have parser ignore standalone whitespace instead of stripping tokens
Summary:
In the current implementation, we modify the `vector<token>` to remove standalone whitespace before passing it to the parser.

In this diff, we change it so that the standalone checker code just marks the position of the tokens and lets the parser logic strip tokens. This cleans up the annoying/hacky token erasure code and bring the core logic back into the `parser` class where it belongs.

I also add `no_parse_result` instead of `std::nullopt_t` to indicate parse failure because we now use `parse_result<std::optional<T>>` so `nullopt` became a bit confusing.

Reviewed By: yoney

Differential Revision: D61883898

fbshipit-source-id: f8c92b93804d0888abc73c837da0ae279039ce8f