Support @rust.Ord structured annotation on unions
Summary:
The Rust Thrift templates already correctly handle unions that need an Ord impl; it just turns into a `#[derive(Ord)]` on the generated data structure.

But unless the implementation of the `rust.Ord` annotation declares that it is compatible with unions, the Thrift compiler rejects `rust.Ord` on unions.

```lang=text,counterexample
[ERROR:dataswarm/services/if/operators.thrift:80] `Ord` cannot annotate `UnionOfAny`
```

Reviewed By: zertosh

Differential Revision: D62074164

fbshipit-source-id: a22908d1b54ba1b436aec7667bbceddc0acf712a