Add support for injecting global names via render_options
Summary:
This diff implements global variables for whisker's rendering engine.

For most cases, this shouldn't be needed. However, this allows `strict_undefined_variables` to remain strict when we integrate with our existing mustache code.

In {D61769182}, we allowlist the subset of variables where our current mustache code relies on being conditionally undefined. This allows the default for new code to error on undefined instead of silently falling back to empty string.

Reviewed By: yoney

Differential Revision: D62050726

fbshipit-source-id: dcc060a4c5dc125007850287f10da33e7adc0142