Extend ::class special handling in DDP to handle nameof
Summary:
In D46593044 and D49833785 there was some special handling added for `::class` expressions for Facts in the direct decl parser. This extends this handling to have matched behavior when Nameof is present in those positions.

Specifically - bare nameof will produce a `Classname` variant, while in a concatenation it will be elaborated and folded (as with ::class).

Reviewed By: nt591

Differential Revision: D62145923

fbshipit-source-id: ccacb31cfa8212b38eb487d1939fa4eadaf91a01