Generalize ApproxSet to Compute More Relations
Summary:
This diff introduces changes to the `ApproxSet` module, which is used to represent sets of elements in an abstract domain. I add a new `ApproxSet.relate` function to compute the relationship between two sets, expanding the capability of the module to reason about more than just disjointness between sets.

The bulk of the logic is contained in `ApproxSet.Set_relation`, where I've added additional functions `complement`, `union` and `inter`. This derives safe assumptions on the relationship between a set, given we know the relation between the other set and its subcomponents.

I utilized an Alloy model to validate that these assumptions are well-formed. I've checked it in, mainly as a reference and don't know if we actually have any mechanisms for running the actual model.

Instead I added a new quick test to `approxSet_test.ml` that tests the well-formed property described in the model.

I plan on making three more changes to this module:
1. Model the universal and empty set
2. Remove the `ApproxSet.diff` function and expose the `ApproxSet.complement` function directly
3. Perform simplifications on the `ApproxSet.t` internal structures

Reviewed By: CatherineGasnier

Differential Revision: D61956448

fbshipit-source-id: be62c4a196cd4815ffe100cef17bcef3d505e242