remove kv_unstable_std feature from log crate
Summary:
Unconditionally enabling this feature on the crate resulted in "std"
features being enabled in the no_std universe.

Note that the feature is still selected by default in the default
universe, so this ought affect only the no_std universe.

Note that this flag was initially enabled in D29804394 and the
author is no longer here to explain the rationale.

Reviewed By: dtolnay

Differential Revision: D62248924

fbshipit-source-id: f0c6ac4dd6bf90be5a84f28cc03e887ae88fb843