Turn on lints if `warnings_default_all = false`
Summary: So far, lints were turned on only if they were off as warnings. That was implemented by checking hack_warnings only. However we've since then added flag `warnings_default_all` which needs to be accounted for as well. This is what this diff does. That requires wiring down `warnings_default_all` to GlobalOption.t

Reviewed By: madgen

Differential Revision: D62178651

fbshipit-source-id: 2efc1a33f7879bcaf6fdb92c8b0f1bcf9fff612e