Hover: show instantiated signature for constructors in generic classes
Summary:
When hovering over a call to a generic method, the user sees the *instantiated* signature.
{F1849312385}
But when hovering over the class in a `new C` expression, the user sees the uninstantiated generic signature.
 {F1849317892}
To be consistent, let's use the instantiated signature, as below.
{F1849313544}
This is a simplification of code in `Ide_hover` at the expense of a small "hack" in `ServerInferTypes`.

Reviewed By: madgen

Differential Revision: D62241073

fbshipit-source-id: 73fb80ba6767a2aef104120c3fd21dd97a4baf4a