Don't report inherited docblocks in origins
Summary:
On hover, we show where the docblock comes from. The implementation naively considers all member decls even if they are inherited. This means we can report a huge number of useless origins when the class has large hierarchies.

This diff changes the implementation so that we only report an origin if the origin definition contains the docblock in its source.

Differential Revision: D62341114

fbshipit-source-id: c3e5520fcf24a8d747395f335540b9054217feb4