Autocomplete with label literals
Summary:
We were autocompleting function/methods calls with variables based on the declaration parameter names. This is almost always the wrong thing when the function/method takes a label.

This diff changes the autocomplete behaviour so that if the parameter type is an enum class label, we create a placeholder label instead.

Differential Revision: D62274184

fbshipit-source-id: c437ce7805a3e9d68b328428f9a3b8e0f1675283