Implement if-blocks
Summary:
This diff implements parsing + rendering of `{{#if}}` blocks from Handlebars:
https://handlebarsjs.com/guide/builtin-helpers.html#if

An if block looks like:
```
{{#if condition}}
  foo
{{else}}
  bar
{{/if}}
```

Most of the complexity in the parser here comes from the back that `{{else}}` is actually a "block" construct even though it looks like variable interpolation. Like other blocks, `{{else}}` is eligible for standalone stripping (which was actually trivial to implement yay!)

Like section blocks, `strict_boolean_conditional` and its coercion rules apply to if blocks as well.

Reviewed By: vitaut

Differential Revision: D62263182

fbshipit-source-id: 13c0450d26c2175f9fedd428644bad8ccfc8c454