OpenSSLHash: Expose hash_size() and block_size()
Summary:
Given an folly::OpenSSLHash::Digest, which can represent any number of message
digests, expose several functions to query properties of the underlying
message digest algorithm. This is useful when the algorithm is given an
arbitrary `Digest` object that was created by a caller, and would like to,
for instance, determine the size of the message digest.

Reviewed By: zalecodez

Differential Revision: D62269695

fbshipit-source-id: 009818bd62d30d31b143ea984bf455b01a7e48ef