Add TupleData and ShapeData to Typing_case_types
Summary:
Inside `Typing_refinement` there is special casing for negations of `IsTupleOf` and `IsShapeOf` predicates. This is because we treat them as `VecData` and `DictData` respectively. I feel this is better handled within `Typing_case_types` and it is a fairly straightforward change to make. We add `TupleData` and `ShapeData` to `Typing_case_types.Tag.t`. We then state that the relationship between `TupleData` and `VecData` is unknown. Since we assume nothing about the relationship between the two data types, we will make the correct conservative conclusion when negating the predicate. The same logic applies for `ShapeData` and `DictData`.

Finally I also state that `TupleData` and `ShapeData` compare as unknown to themselves. This is because as structural types we will compare them by structure when we perform refinement, but we don't represent that in terms of the data type.

Reviewed By: viratyosin

Differential Revision: D62418576

fbshipit-source-id: 1cf3dd8f8d7141b7cf6264d818dcef042b356887