Fix option parsing with decompress-zhhdg
Summary:
This was apparently broken by D62103867, but it's still a bit mysterious how it used to work before that diff. Probably the root was set to a dummy path?

I'm planning to refactor this further next so this cannot happen again with commands we may add in the future.

Differential Revision: D62443050

fbshipit-source-id: 4a7fe0a13a64e91170a732f8f7c4289616fbe359