Migrate existing infer logic to use protocol::Object
Summary:
This diff cribs heavily from D61694974. It implements a version of `infer` (https://fburl.com/code/dvic02s5) without hacking into Python patch internals. This leverages the diff visitor pattern introduced by the Thrift team.

This is as close to the existing logic as I can get. There is one remaining issue which is that our existing infer logic will unconditionally `ensure` fields if they are patched; we do not infer `patchIfSet`s. This will need to be addressed in a followup diff when the pattern is possible (assuming we want to keep current semantics).

Differential Revision: D61944201

fbshipit-source-id: 143c8c63b05a4e991171b94a982828f760092b1e