Fix bug that `remove` operation in nested field is dropped after serialization
Summary:
This is an old bug when we introduce `remove` operation -- we did not check `remove` field when checking `modifies()`.

Later, we use `modifies()` to check whether patch is empty. When serializing patch, since all patch fields are terse, we use the same function to check whether it's empty and then skip the empty field.

Reviewed By: thedavekwon, pranavtbhat

Differential Revision: D62269880

fbshipit-source-id: e9c6d72ca4d3bd96190db7838d01c40b4d4dfddc