log client in query execute
Summary:
As we consider potential future options for Watchman, it's helpful to know
who the watchman clients are. And how many requests they are sending us.

This is the log event for handing a client query. This has a bit more
information about the query than the generic command dispatch log, so let's
include the client information here as well.

Differential Revision: D61723746

fbshipit-source-id: 9b21b2888f24dbce31c4500e6c18ee3781b7d7a7