add ContextConn to thrift servers
Summary:
## TL;DR;
As we can *process* multiple requests over the same connection, it would be nice if we can create an entity and tie it to the lifetime of the connection (this is already implemented in [golang net/http](https://pkg.go.dev/net/http#:~:text=ConnContext%20func(ctx%20context.Context%2C%20c%20net.Conn)%20context.Context)). This will be very beneficial to our team as it allow us to store some state in the server side that we can read/write across multiple requests (in other words, we can identify a client with some server side state, given that this client is using the same connection).

This diff adds `ConContext` to thrift servers, that can modify `ctx` passed to procedures per connection.

Reviewed By: awalterschulze

Differential Revision: D62243055

fbshipit-source-id: ca0eea62f928d75c8bd4c4613d9abdb8ad45d284