hack warnings: Apply error filter in serverless IDE
Summary:
The IDE has two path to get errors: errors from hh_server via the error file, and errors via Tast_provider (that path is part of what we call "serverless IDE").
So far, we were only applying the error filter (`Filter_errors.Filter.t`) in the first path, but not the second. This diff fixes that.

To ensure that as many paths are filtered as possible, we filter errors as close to error production as possible, i.e. in Tast_provider. That means wiring down the `Filter_errors.Filter.t` there from ClientLsp.

Reviewed By: madgen

Differential Revision: D62300829

fbshipit-source-id: a8fb3f1885171ca562c5de8fc1ea54e0348fcac9