prevent exception in ServerRenameTypes
Summary:
I got an exception in ServerRenameTypes.write_patches_to_buffer complaining that the `len` param to `String.sub` was negative. On close examination, there's little guarantee that `i <= len` in that code (see modified line in this diff), so I'm ensuring the `len` argument to `String.sub` is always positive.

Btw, this code is hard to understand, I'm hoping D62391985 and D62394047 on top of this diff make it easier to understand, have a look.

There's still probably a bug because the patches' positions should be smaller than the length of `original_contents` here, but I'm hoping the bug will manifest in odd diffs rather than that undebugable exception.

Reviewed By: madgen

Differential Revision: D62574804

fbshipit-source-id: a5f8d51e477814b2df45b0c9f41369be90948de0