hack warnings: disable individual warnings via .hhconfig flag
Summary:
So far, `hack_warnings` flag could be set to `true`, `false` or a list of authorized codes. However, `hack_warnings` is currently set via JustKnobs, which does not allow to set it to a list of codes. So we set the type of `hack_warnings` back to being a book and add a new .hhconfig flag `disabled_warnings` to allow controlling which codes we want to generate.

I opted for a list of disabled warnings rather than enabled warnings because in the long run, I expect the list of disabled to be shorter than the list of enabled. Plus it makes it easier to add new warning codes.

Differential Revision: D62378235

fbshipit-source-id: 65ba672a229c270cd0a7ec6d7a2ec2d9310d5a9b