Do not show superfluous generic parameters to functions in hover
Summary: Function types as stored in `Call` nodes in the TAST are *instantiated* i.e. generic parameters have been replaced by generic arguments. However, the `ft_tparams` field is not removed; it is still used in some TAST checks. But for hover, it's confusing to see the generic parameters when there are no occurrences in the signature, so to improve the user experience we can remove them before the user sees them in hover and signature help.

Reviewed By: madgen

Differential Revision: D62626594

fbshipit-source-id: 6b8c0b9f8a600a4e677cb93cf948406750dcc241