Show singleton label type in hover
Summary:
For label-based function calls, the label parameter type can be overwhelming and unhelpful to the user:
```
function app<Ta, Tr>(
  HH\EnumClass\Label<Funcs, Func<(int, int), int>> $l,
  (int, int) $arg
): int
```
When hovering over a call to a function whose argument is an explicit label such as `#Mul`, or when showing signature help for such a function whose first argument has already been supplied, it's more friendly to present the singleton type `#Mul` in the signature, even though this isn't strictly the signature of the function itself. We do this for any explicit label argument that is compatible with the specified parameter.
```
function app<Ta, Tr>(#Mul $l, (int, int) $arg): int
```
(Note the presence of the generic parameters: these will be eliminated in a future diff).

Reviewed By: madgen

Differential Revision: D62496510

fbshipit-source-id: c109e6e4a72d6a0879fa195d5a672cfb08c32a33