More Precise Model of Function Types in Typing_case_types
Summary:
# Context
The current model for function types is conservative, essentially treating it as containing the same set of tags as the mixed type. This choice was made because of the various number of data types that the runtime will allow to be invoked and resolve to a function/method. However instead of considering all possible tags the runtime allows to be invoked, we should instead model the set of tags values can have which the type checker considers well typed. It can be broken down into three cases
1. Lambda expressions. These will be instances of the Closure class at runtime. This type is denotable via an hhi
2. Meth_caller expression. This will be an instance of the __SystemLib\MethCallerHelper class. This is not denotable since no hhi exist, so should be considered an implementation detail that isn't exposed
3. Function/Method references. This will be a value with the tag Func or ClsMeth respectively. This is not denotable within the surface language, but we could introduce these tags in our model and have `HH\FunctionRef<>` be the union of these tags

Given this we can consider the set of tags that a value assigned a function type to either be `InstanceOf (Closure)` or `BuiltInData`. This diff adjusts the model within typing_case_types to reflect this.

# Implementation Details
There were unsound assumptions made in the disjoint logic that is within `Typing_subtype`. I updated the logic to no longer consider function types as disjoint from the `Closure` class. Ideally this logic should use the logic within `Typing_case_types`, but I want to wait until the refinement refactor is complete before doing so.

Another change necessary was to adjust the relationship `BuiltInData` has with itself and the `ObjectData` tag. `BuiltInData` is intended to be a catch all for the various tags that a value can have at runtime which is not captured in the model (usually because there isn't a type that denotes it). Given there are also objects that exist at runtime that cannot be denoted, like MethCallerHelper, I've elected to be more conservative in the model. This means that `BuiltInData` is no longer considered equal to itself and is no longer disjoint from `ObjectData`, both are now `Unknown`.

Reviewed By: madgen

Differential Revision: D62615093

fbshipit-source-id: b43eddc5d5bbfdd6efdbf4bfa1bd4299b8102194